feat: added RouteParameterTransformer can convert available args in the router Fixes #16
