+++ packages/IoC/RouterExecutionContext.ts
@@ -4,19 +4,15 @@ import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 import { ContainerMetadata } from "./scanners/MetadataScanner";
 import {
     ClazzType,
-    CustomParamDecoratorMetadata,
     HEADER_TOKEN,
-    HTTP_PARAM_DECORATOR_TOKEN,
     HttpMethod,
-    ServerContext,
     ValidationHandler,
-    createSessionProxy,
-    getParamDecoratorUniqueKey,
 } from "@stingerloom/common";
-import { ValidationError, validate } from "class-validator";
-import { classToPlain, plainToClass } from "class-transformer";
+import { ValidationError } from "class-validator";
+import { classToPlain } from "class-transformer";
 import path from "path";
 import { GuardConsumer } from "./GuardConsumer";
+import { RouteParameterTransformer } from "./RouteParameterTransformer";
 
 /**
  * @class RouterExecutionContext
@@ -64,48 +60,17 @@ export class RouterExecutionContext {
                         routerName,
                     );
 
+                    const routeParameterTransformer =
+                        new RouteParameterTransformer(
+                            req,
+                            targetController,
+                            routerName,
+                            bodyValidationActions,
+                        );
+
                     // Parameters Consumer
-                    const args = parameters.map((param) => {
-                        // Req 객체 요청
-                        if (param.isReq) {
-                            return req;
-                        }
-
-                        // 세션 요청
-                        if (param.isSession) {
-                            return createSessionProxy(req);
-                        }
-
-                        // 커스텀 매개변수 구현
-                        if (param.isCustom) {
-                            const callback = Reflect.getMetadata(
-                                HTTP_PARAM_DECORATOR_TOKEN,
-                                targetController,
-                                routerName,
-                            ) as CustomParamDecoratorMetadata;
-
-                            const context = new ServerContext(req);
-                            return callback[
-                                getParamDecoratorUniqueKey(
-                                    targetController,
-                                    routerName,
-                                    param.index,
-                                )
-                            ].callback(param.value, context);
-                        }
-
-                        // Body 요청 구현
-                        if (param.body) {
-                            const bodyData = plainToClass(
-                                param.body.type,
-                                req.body,
-                            );
-                            bodyValidationActions.push(validate(bodyData));
-                            return bodyData;
-                        }
-
-                        return param.value;
-                    });
+                    const args =
+                        routeParameterTransformer.transform(parameters);
 
                     // Header Consumer
                     const header = Reflect.getMetadata(
