+++ packages/IoC/index.ts
@@ -3,3 +3,4 @@ export * from "./ContainerManager";
 export * from "./JSDocReader";
 export * from "./RouterExecutionContext";
 export * from "./GuardConsumer";
+export * from "./RouteParameterTransformer";
